integated audio section player changes into audio player widget
