+++ lib/pages/chat/events/audio_player.dart
@@ -25,7 +25,13 @@ class AudioPlayerWidget extends StatefulWidget {
 
   static String? currentId;
 
-  static const int wavesCount = 40;
+  // #Pangea
+  // static const int wavesCount = 40;
+  static const int wavesCount = 100;
+
+  final double? sectionStartMS;
+  final double? sectionEndMS;
+  // Pangea#
 
   const AudioPlayerWidget(
     this.event, {
@@ -33,6 +39,8 @@ class AudioPlayerWidget extends StatefulWidget {
     // #Pangea
     this.matrixFile,
     this.autoplay = false,
+    this.sectionStartMS,
+    this.sectionEndMS,
     // Pangea#
     super.key,
   });
@@ -59,6 +67,20 @@ class AudioPlayerState extends State<AudioPlayerWidget> {
   MatrixFile? matrixFile;
   File? audioFile;
 
+  // #Pangea
+  int? get startMS {
+    return widget.sectionStartMS != null
+        ? (widget.sectionStartMS! * maxPosition).round()
+        : null;
+  }
+
+  int? get endMS {
+    return widget.sectionEndMS != null
+        ? (widget.sectionEndMS! * maxPosition).round()
+        : null;
+  }
+  // Pangea#
+
   @override
   void dispose() {
     if (audioPlayer?.playerState.playing == true) {
@@ -71,6 +93,24 @@ class AudioPlayerState extends State<AudioPlayerWidget> {
 
     super.dispose();
   }
+
+  // #Pangea
+  @override
+  void didUpdateWidget(covariant oldWidget) {
+    if ((oldWidget.sectionEndMS != widget.sectionEndMS) ||
+        (oldWidget.sectionStartMS != widget.sectionStartMS)) {
+      debugPrint('selection changed');
+      if (startMS != null) {
+        audioPlayer?.seek(Duration(milliseconds: startMS!));
+        audioPlayer?.play();
+      } else {
+        audioPlayer?.stop();
+        audioPlayer?.seek(null);
+      }
+    }
+    super.didUpdateWidget(oldWidget);
+  }
+  // Pangea#
 
   Future<void> _downloadAction() async {
     // #Pangea
@@ -160,7 +200,15 @@ class AudioPlayerState extends State<AudioPlayerWidget> {
                 AudioPlayerWidget.wavesCount)
             .round();
       });
-      if (state.inMilliseconds.toDouble() == maxPosition) {
+      // #Pangea
+      if (startMS != null &&
+          endMS != null &&
+          state.inMilliseconds.toDouble() >= endMS!) {
+        audioPlayer.stop();
+        audioPlayer.seek(Duration(milliseconds: startMS!));
+      } else if (state.inMilliseconds.toDouble() == maxPosition) {
+        // if (state.inMilliseconds.toDouble() == maxPosition) {
+        // Pangea#
         audioPlayer.stop();
         audioPlayer.seek(null);
       }
@@ -194,6 +242,11 @@ class AudioPlayerState extends State<AudioPlayerWidget> {
       }
       // Pangea#
     }
+    // #Pangea
+    if (startMS != null) {
+      audioPlayer.seek(Duration(milliseconds: startMS!));
+    }
+    // Pangea#
     audioPlayer.play().onError(
           ErrorReporter(context, 'Unable to play audio message')
               .onErrorCallback,
@@ -311,6 +364,16 @@ class AudioPlayerState extends State<AudioPlayerWidget> {
 
     final statusText = this.statusText ??= _durationString ?? '00:00';
     final audioPlayer = this.audioPlayer;
+
+    // #Pangea
+    final msPerWave = (maxPosition / AudioPlayerWidget.wavesCount);
+    final int? startWave = startMS != null && msPerWave > 0
+        ? (startMS! / msPerWave).floor()
+        : null;
+    final int? endWave =
+        endMS != null && msPerWave > 0 ? (endMS! / msPerWave).ceil() : null;
+    // Pangea#
+
     return Padding(
       // #Pangea
       // padding: const EdgeInsets.all(12.0),
@@ -352,44 +415,122 @@ class AudioPlayerState extends State<AudioPlayerWidget> {
           // #Pangea
           // const SizedBox(width: 8),
           const SizedBox(width: 5),
-          // Pangea#
-          Row(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              for (var i = 0; i < AudioPlayerWidget.wavesCount; i++)
-                GestureDetector(
-                  onTapDown: (_) => audioPlayer?.seek(
-                    Duration(
-                      milliseconds:
-                          (maxPosition / AudioPlayerWidget.wavesCount).round() *
-                              i,
-                    ),
-                  ),
-                  child: Container(
-                    height: 32,
-                    color: widget.color.withAlpha(0),
-                    alignment: Alignment.center,
-                    child: Opacity(
-                      opacity: currentPosition > i ? 1 : 0.5,
-                      child: Container(
-                        margin: const EdgeInsets.symmetric(horizontal: 1),
-                        decoration: BoxDecoration(
-                          color: widget.color,
-                          borderRadius: BorderRadius.circular(2),
-                        ),
-                        // #Pangea
-                        // width: 2,
-                        width: 1,
-                        // Pangea#
-                        height: 32 * (waveform[i] / 1024),
-                      ),
-                    ),
-                  ),
-                ),
-            ],
-          ),
-          // #Pangea
+          // Row(
+          //   mainAxisSize: MainAxisSize.min,
+          //   children: [
+          //     for (var i = 0; i < AudioPlayerWidget.wavesCount; i++)
+          //       GestureDetector(
+          //         onTapDown: (_) => audioPlayer?.seek(
+          //           Duration(
+          //             milliseconds:
+          //                 (maxPosition / AudioPlayerWidget.wavesCount).round() *
+          //                     i,
+          //           ),
+          //         ),
+          //         child: Container(
+          //           height: 32,
+          //           color: widget.color.withAlpha(0),
+          //           alignment: Alignment.center,
+          //           child: Opacity(
+          //             opacity: currentPosition > i ? 1 : 0.5,
+          //             child: Container(
+          //               margin: const EdgeInsets.symmetric(horizontal: 1),
+          //               decoration: BoxDecoration(
+          //                 color: widget.color,
+          //                 borderRadius: BorderRadius.circular(2),
+          //               ),
+          //               // #Pangea
+          //               // width: 2,
+          //               width: 1,
+          //               // Pangea#
+          //               height: 32 * (waveform[i] / 1024),
+          //             ),
+          //           ),
+          //         ),
+          //       ),
+          //   ],
+          // ),
           // const SizedBox(width: 8),
+          Expanded(
+            child: Row(
+              children: [
+                for (var i = 0; i < AudioPlayerWidget.wavesCount; i++)
+                  Builder(
+                    builder: (context) {
+                      final bool hasSelection =
+                          endMS != null && startMS != null;
+
+                      final bool inRange = startWave != null &&
+                          i >= startWave &&
+                          endWave != null &&
+                          i < endWave;
+
+                      double barOpacity = currentPosition > i ? 1 : 0.5;
+                      if (hasSelection && !inRange) {
+                        barOpacity = 0.5;
+                      }
+
+                      return Expanded(
+                        child: Stack(
+                          children: [
+                            Expanded(
+                              child: GestureDetector(
+                                onTapDown: (_) => audioPlayer?.seek(
+                                  Duration(
+                                    milliseconds: (maxPosition /
+                                                AudioPlayerWidget.wavesCount)
+                                            .round() *
+                                        i,
+                                  ),
+                                ),
+                                child: Expanded(
+                                  child: Container(
+                                    height: 32,
+                                    color: widget.color.withAlpha(0),
+                                    alignment: Alignment.center,
+                                    child: Opacity(
+                                      opacity: barOpacity,
+                                      child: Expanded(
+                                        child: Container(
+                                          margin: const EdgeInsets.symmetric(
+                                            horizontal: 1,
+                                          ),
+                                          decoration: BoxDecoration(
+                                            color: widget.color,
+                                            borderRadius:
+                                                BorderRadius.circular(2),
+                                          ),
+                                          height: 32 * (waveform[i] / 1024),
+                                        ),
+                                      ),
+                                    ),
+                                  ),
+                                ),
+                              ),
+                            ),
+                            if (startWave != null &&
+                                i >= startWave &&
+                                endWave != null &&
+                                i < endWave)
+                              Expanded(
+                                child: Opacity(
+                                  opacity: 0.5,
+                                  child: Container(
+                                    height: 32,
+                                    decoration: const BoxDecoration(
+                                      color: Colors.green,
+                                    ),
+                                  ),
+                                ),
+                              ),
+                          ],
+                        ),
+                      );
+                    },
+                  ),
+              ],
+            ),
+          ),
           const SizedBox(width: 5),
           // SizedBox(
           //   width: 36,
