+++ lib/pangea/widgets/chat/message_audio_card.dart
@@ -1,9 +1,9 @@
 import 'dart:math';
 
+import 'package:fluffychat/pages/chat/events/audio_player.dart';
 import 'package:fluffychat/pangea/matrix_event_wrappers/pangea_message_event.dart';
 import 'package:fluffychat/pangea/models/pangea_token_model.dart';
 import 'package:fluffychat/pangea/utils/error_handler.dart';
-import 'package:fluffychat/pangea/widgets/chat/audio_section_player.dart';
 import 'package:fluffychat/pangea/widgets/chat/message_selection_overlay.dart';
 import 'package:fluffychat/pangea/widgets/chat/message_toolbar.dart';
 import 'package:fluffychat/pangea/widgets/chat/toolbar_content_loading_indicator.dart';
@@ -133,11 +133,12 @@ class MessageAudioCardState extends State<MessageAudioCard> {
           : localAudioEvent != null || audioFile != null
               ? Column(
                   children: [
-                    AudioSectionPlayerWidget(
+                    AudioPlayerWidget(
                       localAudioEvent,
                       matrixFile: audioFile,
                       sectionStartMS: sectionStartMS,
                       sectionEndMS: sectionEndMS,
+                      color: Theme.of(context).colorScheme.onPrimaryContainer,
                     ),
                   ],
                 )
